feat/password-hashing: move functions over to model.go for better integration
