+++ api/api/api.go
@@ -3,7 +3,6 @@ package api
 import (
 	"api/model"
 	"api/storage"
-	"crypto/rand"
 	"encoding/json"
 	"log"
 	"net/http"
@@ -208,16 +207,6 @@ func hashPassword(password string) string {
 	return hashedPassword
 }
 
-func randomSalt(length uint32) ([]byte, error) {
-	secret := make([]byte, length)
-
-	_, err := rand.Read(secret)
-	if err != nil {
-		return nil, err
-	}
-
-	return secret, nil
-}
 
 func (s *APIServer) wrapInJSON(objects ...interface{}) (string, error) {
 	jsonData := make(map[string]interface{})
